+++ crates/resvg/src/image.rs
@@ -18,7 +18,7 @@ pub struct Image {
 }
 
 pub fn convert(image: &usvg::Image, children: &mut Vec<Node>) -> Option<BBoxes> {
-    let object_bbox = image.view_box.rect.to_rect();
+    let object_bbox = image.bounding_box?.to_rect();
     let bboxes = BBoxes {
         object: usvg::BBox::from(object_bbox),
         layer: usvg::BBox::from(object_bbox),